refactor: merge integration feature with my
