+++ expo/app/(tabs)/profile/account/edit.tsx
@@ -2,20 +2,20 @@ import { useEffect, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { ScrollView, StyleSheet } from "react-native";
 
-import Alert from "../../components/Alert";
-import Button from "../../components/Button";
-import ControlledInput from "../../components/ControlledInput";
-import { View } from "../../components/Themed";
-import Warning from "../../components/Warning";
-import Avatar from "../../components/profile/Avatar";
+import Alert from "../../../../components/Alert";
+import Button from "../../../../components/Button";
+import ControlledInput from "../../../../components/ControlledInput";
+import { View } from "../../../../components/Themed";
+import Warning from "../../../../components/Warning";
+import Avatar from "../../../../components/profile/Avatar";
 import {
   AuthErrorMessage,
   SupabaseErrorCode,
-} from "../../constants/SupabaseErrorCode";
-import { emailRules, usernameRules } from "../../lib/inputRestriction";
-import { supabase } from "../../lib/supabase";
-import { useSupabaseUserHook } from "../../lib/useSupabaseUser";
-import { getUserProfile } from "../../lib/userProfile";
+} from "../../../../constants/SupabaseErrorCode";
+import { emailRules, usernameRules } from "../../../../lib/inputRestriction";
+import { supabase } from "../../../../lib/supabase";
+import { useSupabaseUserHook } from "../../../../lib/useSupabaseUser";
+import { getUserProfile } from "../../../../lib/userProfile";
 
 type EditForm = {
   email: string;
